requirement.txt
numpy==1.26.4
opencv-python==4.8.1.78
faiss-cpu==1.7.4
insightface==0.7.3
onnxruntime==1.20.0
ultralytics==8.1.28
scikit-learn
